Move old path types to index
